Move device ID actions to a properly-named file
